log: add edge function checking
Made use of edge functions which is essentially the same as my previous
method of checking within bounds except computed differently and
provides better support for tile-based rasterizing.

Also added edge stepping so that the edge functions don't need to be
recomputed for each pixel, this is shown through the difference b/w
renderMesh2 and renderMesh3.