log: finish graphics renderer
After some careful consideration, I've decided to pull the plug on this.
I'm spending way too much time debugging and trying to understand
mathematical concepts (even if they arent that hard) instead of actually
coding anything. My inital goal was to make a graphics renderer to be
used within a game engine, but at this rate that's never going to
happen.

Also, after the bomb that was tile based rasterization I realized that I
would need to make tons of tiny optimizations to squeeze as much
performance out of this as possible.

This is the end goodbye!